pyproject.toml
[tool.poetry]
name = "dev_ops"
version = "0.1.0"
description = ""
authors = ["algernon"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.1.5"
mysqlclient = "2.0.1"
django-pure-pagination = "^0.3.0"
yapf = "^0.30.0"
django-redis = "^4.12.1"
ansible = "^2.10.5"
paramiko = "^2.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "qinghua"
url = "https://mirrors.aliyun.com/pypi/simple/"
